// packages/collection-store/src/auth/rbac/types.ts
export const SYSTEM_ROLES = {
  SUPER_ADMIN: 'system:super_admin',
  ADMIN: 'system:admin',
  USER: 'system:user',
  GUEST: 'system:guest',
  SERVICE: 'system:service'
} as const